Some work on generating genomes
Addes support for the graph returning an array with the IDs of all
the required elements, shuffeling integer arrays, generating a
random genome, and making the genomes comparable on fitness. Also
made space for the testing methods for the code in the evolutionary
algorithm part of the program.

A thing that has come to mid would be having the graph being a
static instance. We do a lot of accessing and passing around of
something that essentially is an unalterable singleton, so this
could impact performance as well as general neatness.