Added adult selection and fixed sorting
The genotypes were sorted from low to hight, but reveresd it to
high to low. This is because in our case low is better than high,
and some routines (in selection especially) rely on that the best
individuals are at the end of lists.

Also added adult selection and lots of comments, probably also some
general improvements.