EA fixed
The error was something regarding the resetFitness() function, so
I removed it seeing as it is not really critical. Have discovered
a new interesting bug, which is that split doesn't seem to have
all of the required graph elements upon exit. Weird, but not something
that I will fix or even look into at this hour. For now what was
supercritical is done, and I will push it for the sake of its preservation.