Updated formating of output
The outputed trips are now in terms of the names of the graph-
elements that were given as input, instead of the internally used
indexes. Also added option to output as tab-separated data if one
wishes to copy-paste it directly into spreadsheets, but is not used
at the moment and most likely won't ever be.