Fixes tournament selection, but not permanentely
The good part: tournament select no longer breaks because it runs
out of population to have tournaments among.

The bad part: tournament select doesn't work when one does selection
without replacement. So, for it to work as intended, we need to do
replacement while checking that no two conescutive elements in
selected parents are equal. Will probably be messy and add complexity,
and the performance is already not ideal for this kind of selection