More than halved disk size of results
By removing the best genomes of each stored generation. These were
more than half of each line in the files, and absolutetly useless
for the purposes we have for the circle test in our paper. Sure, on
a real life case, it might be interesting to see how the best solution
changes over time on a map, but we have no utility for displaying
the change of the circle problem each generation (we would have to
make our own map or find a graph visualization tool that works with
the graph format sintef has chosen to use for these graphs). Also
because we keep the seed the entire process can be exactly reproduced,
so in the unlikely event that someone wants to look at those, they
can download them themselves.

Also added some more graph pdfs, which fortunately indicate that I
still make sane conjectures (a larger population size is better in
the case of fitness proportionate selection, actually probably in
most realistic cases other than uniform selection full replacement).

The graph pdfs should however all be reworked to be exactly the same
scale (at the moment they are somewhat unike snowflakes), and we should
look into making the pdfs so that they are exactly the sie of the
graph (ie. not a print on a page with a lot of additional whitespace)...