Completed the mutation graphs
Now only to include them in the latex. The conclusions should be
changed to indicate that adding the mutated child gives a slightly
faster improvement in the beginning, but means jack shit in the long
run and therefore shouldn't be used because it uses more resources.

Also, we can see that memetic improvement decreases the diversity of
the population, but not enough to be bad (think of the elitist mixing
graph).

(And the circle test figures should be divided into subfolders, the
one that's currently in use is starting to get messy...)