Some folder reorganizing and work on the EAChap
The figures are now much more neatly arranged, and the EA tuning
text is much better now that it has been sprinkled with lots of
references to the charts straight next to the references..

Oh, and I also added a url-library so that the links get properly
line-broken and actually link to web-sites if they are clicked!
Should probably look into whether the styling should be the same
as for the rest of the text, but for now that is less important than
beginning to write *cough*paraphrase from the pre-project*cough* the
theoretical background chapter.

But now more coffee and writing job-applications ^^