Model desctiptions decisions (almost?) complete
One should perhaps give a full formal definition of the NEARP that
will be used for the test cases and internal structure at this point.
Then one should begin writing up what approaches have been used, and
what approach we chose and why. Thereafter Begins the general theory
behind the algorithms that are to be implemented. Then the implementation
and methodology chapter.