Completed NEARP section, begun routing algorithms
The nearp section is still sans sources, and the beginning of the
routing algoritm section has to be determined to either lay out
the historical approach, or the "these are the different types"
approach (which will still be sort of a historical ordering).

History flows well, but the "by type"-approac has a nice feeling of
to the point and concise. I guess we'll see what comes out ^^