Some work on the EA introduction section
Also fixes of minor typos and latex-badness. At this point a smooth
transition to talking about what an MA is for the snow plowing problem,
and going into details about the fitness for snow plowing. What we
actually implement should perhaps be in mathod and implementation,
and what we don't implement in future work?