Some more work on the EA theory
What should be done now is that a transition to discussing details
of what is theoretically needed of a MA that we can implement, and
write that part (heavily relying on the pre-made pseudocode, and
major resources should be expended to make a good explanation of
split). Then the methodology and implementation can be forgone for
a while untill some wild test results appear in form of emails from
the municipality, and in the mean time the wording of the conclusions
in the EA-tuning chapter can be imroved.

After that there is only methodology, actually obtaining results to
analyze, and sprinkeling the whole thing with sources that remains?