EA tuning chapter much improved
Not lists a plausible reason for why the small population result
was obtained. The split thing needs working, for now try to say
something about that it works by looking at each required element,
sees if it can remove the retrun to the depto and replace it with one
more required element and then returning to the depot. Reasons that
might fails is that by adding one more required task the vehicles
demand capacity could be exeeded, or that the required item is already
part of another splitted trip with an earlier starting point and
that has a lower cost in that trip.