Work on explaining the split algorithm.
Its almost there. It needs reviewing for heavy sentences, transitions
between paragraphs, whether the thing as a whole makes sense, whether
it needs more diagrams (whether to shamelessly rip of ulusoys work
from the 80-s in a new and shiny package so that it can be refferenced
in the "reproduced according to [x paper that is based on his work
but has the more modern readable graph]" way).

Othervise soon ready for the "theory chapter done"-milestone. Shoul
go over the last EA-part to make sure that it is not too focused on
the implementation details. Also should see whether it should be
mentioned that EAs/GAs gusually are local search-thingies...