8000 words milestone
Also I think I've got the external resources used in the input
phase/before the MA sorted out. What remains now is to comment on
how the municipalitys data is combined with the NRDB data into the
NEARP format, and begin discussing the MA implementation details
(Loops, focused areas of the fitness, and so on and so forth...).

After that the only remaining to discuss in this chapter are the
output (simple enough, map with arrows on lines that are route,
however with some possible neat details in intersections that don't
have to be discussed at this point in time). Then transition into
the MA-tuning chapter. (From which one goes to the actuall tests
and results, to the evaluation and conclusion, which ends in the
future work.)

One issue will be that Viken is not truely available untill June
the 7th, and there still remains data to be collected from the
municipality (concre routes being driven by drivers currently).
That might be a good time to write a good speck for what I actually
need from Viken and putting stuff in the future work section that
I already now know that needs to be there. Also sources. Don't
forget that you have to do sources.