Parent and addult selection discussed
Also fixed minor typos elsewhere. The next things are a part on the
asympthotical performance, an introduction to snow plowing in the
background section, moving the pseudocode to the appendix, reworking
the introduction to be more to the point and with shiny research-
questions, fixing up the background chapter according to feedback,
more shiny figures, adding a table of the found parameters at the
end of the tuning part, addressing that it addresses a research-
question, adding sources, rewriting the preface, and beginning the
speck for road-extraction stuff.