Made the csv-dat conversion program in python
It converts csv to dat in a most satisfying manner. One should perhaps
consider making a tool that reverses the output to something that
can be used by the system that generated the csv, but for now that
is future work... (although I promised the municipality-dudes that
they would get shiny maps with routes. Some day perhaps..)