Some work and images on snow plowing
All images aproved and/or CC-licensed. The main challenges image-
wise from here are determining how many illustrations of tools to
include, and where to include the muncipal maps. They are partially
relevant for the theory section about what the municipality wants
done, they are also relevant for the implementation part (to show
what was worked with), the should definitely be considered relevant
for the data results/analysis (showing what the results are about/
what the analysis is saying), and for the evaluation (showing what
details affected what parts of the results, and what details are
important for further work on and improving the model).

Of the places mentioned above, I feel that the perhaps are best shown
in the results and analysis section. There they can be directly
related to the output, and they can also be shown side by side with
how the drivers choose to drive them, and what the program has
outputted. Then highlighed parts can be used in the conclusion/
evaluation if that should proove to be desirable.

Regarding the snow plowing section, I think that it should perhaps
be approached in a different way. It might flow better if one starts
out with winter road maintenance, first covering the why (lots of
snow -> society needing to function), then covering the various hows
(gritting, salting, plowing), then going in depth on plowing. Once
that point has been reached, one can discuss some intricate domain
details that are of mostly theoretical importance.

In this case they (the details) should mainly be about the municipalitys
resources and goals. One perhaps starts out with the divide between
them and the NPRA. Then one comes to the size of their problem, and
how they have chosen to divide it into smaller sub-problems, mostly
based on what equipment is best suited for said sub-problems (ie.
large vehicles for large roads, smaller vehicles for smaller roads).
Then it is natural to discuss the different classes of vehicles that
they use for what. Then one mentions that they have a structure where
they have some of the equipment and workers themselves, while at the
same time much of the work is given out to subcontractors that have
their own equipment. The subcontractors range from specialized companies
with several vehicles and drivers, to independent contractos that
own their own vehicle (typically tractors), and operate alone when
they are called in after a snowfall.

Then one finishes of with what factors the municipality might desire
to optimize for. The list should include driving time, distance driven,
vehicle durability, infrastructure durability, fuel consumption,
environmental factors, how quickly the most important objects get
cleared, etc.

This should suffice for an introduction. More intricate details that
might affect choices and outcomes, such as appendixes of one way roads,
the intircacy of intersections, time of day issues, etc., should be
left for the results and analysis, and conclusion and evaluation.