A lot of fixing of stuff, and some additions
First and foremost, alot of the legacy problems with "we" and "our"
in the EA tuning chapter were fixed. One should probably look into
being more consequent with EA vs MA in that chapter, and the use of
past and future tense (this could apply to all of the text actually).

Some maps and figures were added, and a last todo was addressed. It
shpuld also be noted that the generated maps are now added in the
latex part of the project, although they are not used yet. They
should probably be used in the experiments and results sections,
perhaps be referenced in the conclusion chapter, and be considered
for front page graphics.