Some more work on the experiments and results
The BHW1-thingy is more or less done. Did some table-styling with
booktabs (seems to make horizontal lines work better with pdf-readers).
The next big thing should be finishing the drawing of the graphs for
BHW1, and commencing the work on the data from Trondheim-chapter.

When those are done, look into going over some of the text with
grammarly, and what to put in the conclusion and evaluation capter?