Some work on experiments and results
Added references to the tiles with seemingly broken connections in
the drawn maps. They have obviously been lost in the visible map
data, but it is due to the reading in thing where all elements closer
than whatever are merged and things with no speed limmit are removed.

Could also be a simple error in the underlying map data, but that
the route has managed to use the connections even if they are not
shown disprooves this and shows that something is bad in the way
in which they are drawn...