Some work on the conclusion and evaluation
As well as fixing the indexing of the stuff that comes before page
1. The ToC is correct now. Now to figure out what goes in the
evaluation section. The research questions and goal should be re-
stated and answered. But is there anything else? Like that I cannot
find the optimal solutions? Should I say anything about the work
with the map stuff here? Stuff will be figured out in the hopefully
near future...