Major fixings
The pre project is now included in the appendinx, so that it exists.
The messed up figure (3.2) is now moved up on the page so that it
doesn't completely break that page transition. The copytight-
attriburtion texts have been updated so that they are less verbose
and more in tune with what Jo likes. Now for the rest of the fixes.