Did stuff with headings
The first entries (abstract, norwegian abstract, and preface) and
the appendix title now no longer have bars along the head of their
pages. This is to make them look more akin to the chapter title
pages.

The chapter title pages have been moved up, so that the top of the
title bar-figure-drawing-thingy is align with the top of the pages
otherwise. Will have to think about this for a while. It makes the
pagecont somewhat lower. But I think it looks kind of better.

Also chapter 1 still should be considered, but I honestly have no
idea what more to do with it at this point. Maybe something will
come to me after dinner?